requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0.post1
celery==5.3.4
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
duckdb==0.9.2
langchain==0.1.5
faiss-cpu==1.7.4
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dateutil==2.8.2
openpyxl==3.1.2

# Frontend dependencies (managed via npm)
# react, antd, axios, echarts